+++ src/xlnstorch/operators/arithmetic_ops.py
@@ -3,7 +3,6 @@ from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implem
 from ..autograd import LNSFunction
 from . import (
     lns_add,
-    lns_neg,
     lns_mul,
     lns_div,
     lns_square,
@@ -12,222 +11,6 @@ from . import (
     lns_matmul,
     lns_sum,
 )
-
-# SBDB_FUNCS is a dictionary that contains different implementations
-# of the sbdb (Gaussian logarithm) function. Each implementation is
-# registered with a unique key.
-SBDB_FUNCS = {}
-DEFAULT_SBDB_FUNC = ""
-
-def implement_sbdb(key, default=False):
-    """
-    A decorator to register a custom sbdb implementation. This will
-    be used to compute/approximate the Gaussian logarithms for the
-    addition and subtraction operations in the logarithmic domain. See
-
-    https://en.wikipedia.org/wiki/Logarithmic_number_system
-    https://en.wikipedia.org/wiki/Gaussian_logarithm
-
-    Parameters
-    ----------
-    key : str
-        The key to register the sbdb function under. This should be
-        unique across all sbdb implementations.
-    default : bool, optional
-        If True, this sbdb function will be set as the default sbdb
-        implementation. If multiple sbdb functions are registered
-        with `default=True`, the last one registered will be used as
-        the default. Defaults to False.
-
-    Raises
-    ------
-    ValueError
-        If an sbdb function with the given key is already registered.
-    """
-    def decorator(func):
-        function_key = key or func.__name__
-
-        if function_key in SBDB_FUNCS:
-            raise ValueError(f"sbdb function with key '{function_key}' is already implemented.")
-        SBDB_FUNCS[function_key] = func
-
-        if default:
-            global DEFAULT_SBDB_FUNC
-            DEFAULT_SBDB_FUNC = function_key
-
-        return func
-    return decorator
-
-def sbdb(z, s, base):
-    """
-    Computes the Gaussian logarithm for the given inputs z and s.
-
-    Parameters
-    ----------
-    z : torch.Tensor
-        The negation of the absolute difference between the two operands
-        in the logarithmic domain.
-    s : torch.Tensor
-        The sign difference between the two operands in the logarithmic
-        domain.
-    base : torch.Tensor
-        The base of the operands. Required for certain sbdb implementations.
-
-    Returns
-    -------
-    torch.Tensor
-        The result of the Gaussian logarithm computation.
-
-    Raises
-    ------
-    ValueError
-        If no default sbdb function is implemented.
-    """
-    if DEFAULT_SBDB_FUNC not in SBDB_FUNCS:
-        raise ValueError(f"No default sbdb function implemented.")
-
-    return SBDB_FUNCS[DEFAULT_SBDB_FUNC](z, s, base)
-
-@implement_sbdb('ideal', default=True)
-def sbdb_ideal(z, s, base):
-    """
-    Ideal implementation of the sbdb function that directly computes:
-    log_(base)(1 - 2 * s + base ^ z)
-    """
-    power_term = torch.pow(base, z)
-    magnitude = torch.abs(1.0 - 2.0 * s + power_term)
-
-    log_term = torch.log(magnitude) / torch.log(base)
-    result = torch.round(log_term) * 2
-
-    return result.to(torch.float64)
-
-class LNSAddFunction(LNSFunction):
-    """
-    Addition is far more challenging in the logarithmic domain.
-    We can implement different approximate methods for the sum
-    and difference functions (Gaussian logarithms). See
-
-    https://en.wikipedia.org/wiki/Logarithmic_number_system
-    https://en.wikipedia.org/wiki/Gaussian_logarithm
-
-    For two internal representations x and y, their addition can
-    be computed as follows:
-    max(x, y) + sbdb(-|(x >> 1) - (y >> 1)|, (x ^ y) & 1)
-
-    Gradients are computed as follows:
-    d/dx(x + y) = 1
-    d/dy(x + y) = 1
-    """
-
-    @staticmethod
-    def forward(x, y, base):
-
-        x_packed, y_packed = x.to(torch.int64), y.to(torch.int64)
-        max_operand = torch.max(x_packed, y_packed)
-
-        abs_diff = torch.abs((x_packed >> 1) - (y_packed >> 1))
-        sign_diff = (x_packed ^ y_packed) & 1
-
-        result = max_operand + sbdb(-abs_diff, sign_diff, base)
-        return torch.where(
-            torch.eq(x_packed | 1, LNS_ZERO), y, torch.where(
-                torch.eq(y_packed | 1, LNS_ZERO), x, torch.where(
-                    x_packed ^ 1 == y_packed, LNS_ZERO, result.to(torch.float64))))
-
-    @staticmethod
-    def setup_context(ctx, inputs, output):
-        pass # no context needed for this operation
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        return grad_output, grad_output, None
-
-@implements(torch.add, LNSAddFunction.forward, key='default', default=True)
-def add(x, y, *, alpha=1, out=None):
-
-    x, y = format_lnstensor_operands(x, y)
-
-    if alpha != 1:
-        y = torch.mul(y, alpha)
-
-    result = LNSAddFunction.apply(x, y, x.base)
-
-    if out is not None:
-        return out._inplace_copy(result)
-
-    return lnstensor(result, from_lns=True, b=x.base)
-
-class LNSSubFunction(LNSFunction):
-    """
-    See LNSAddFunction for details on the internal computations.
-
-    Gradients are computed as follows:
-    d/dx(x - y) = 1
-    d/dy(x - y) = -1
-    """
-
-    @staticmethod
-    def forward(x, y, base):
-        neg_y = lns_neg(y)
-        return lns_add(x, neg_y, base)
-
-    @staticmethod
-    def setup_context(ctx, inputs, output):
-        pass # no context needed for this operation
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        grad_y = lns_neg(grad_output)
-        return grad_output, grad_y, None
-
-@implements(torch.sub, LNSSubFunction.forward, key="default", default=True)
-def sub(x, y, *, alpha=1, out=None):
-
-    x, y = format_lnstensor_operands(x, y)
-
-    if alpha != 1:
-        y = torch.mul(y, alpha)
-
-    result = LNSSubFunction.apply(x, y, x.base)
-
-    if out is not None:
-        return out._inplace_copy(result)
-
-    return lnstensor(result, from_lns=True, b=x.base)
-
-class LNSNegFunction(LNSFunction):
-    """
-    Negation becomes flipping the sign bit.
-
-    Gradients are computed as follows:
-    d/dx(-x) = -1
-    """
-
-    @staticmethod
-    def forward(x):
-        x_packed = x.to(torch.int64)
-        neg_x_packed = x_packed ^ 1
-
-        return neg_x_packed.to(torch.float64)
-
-    @staticmethod
-    def setup_context(ctx, inputs, output):
-        pass # no context needed for this operation
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        return lns_neg(grad_output)
-
-@implements(torch.neg, LNSNegFunction.forward, key="default", default=True)
-def neg(x, *, out=None):
-
-    result = LNSNegFunction.apply(x)
-
-    if out is not None:
-        return out._inplace_copy(result)
-
-    return lnstensor(result, from_lns=True, b=x.base)
 
 class LNSMulFunction(LNSFunction):
     """
@@ -558,110 +341,6 @@ class LNSLogFunction(LNSFunction):
 def log(x, *, out=None):
 
     result = LNSLogFunction.apply(x, x.base)
-
-    if out is not None:
-        return out._inplace_copy(result)
-
-    return lnstensor(result, from_lns=True, b=x.base)
-
-class LNSAbsFunction(LNSFunction):
-    """
-    Absolute value becomes setting the sign bit off.
-
-    Gradients are computed as follows:
-    d/dx(|x|) = 1 if x > 0, -1 if x < 0 
-
-    Note that PyTorch defines the gradient to be 0
-    when x=0 despite it being undefined here.
-    """
-
-    @staticmethod
-    def forward(x):
-        x_packed = x.to(torch.int64)
-        x_packed_abs = x_packed & (~1)
-
-        return torch.where(torch.eq(x_packed | 1, LNS_ZERO), LNS_ZERO, x_packed_abs.to(torch.float64))
-
-    @staticmethod
-    def setup_context(ctx, inputs, output):
-        x, = inputs
-        ctx.save_for_backward(x)
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        x, = ctx.saved_tensors
-        x_packed = x.to(torch.int64)
-        x_packed_sign = x_packed & 1
-
-        return torch.where(torch.eq(x_packed_sign, 1), lns_neg(grad_output), grad_output)
-
-@implements(torch.abs, LNSAbsFunction.forward, "default", default=True)
-def abs(x, *, out=None):
-
-    result = LNSAbsFunction.apply(x)
-
-    if out is not None:
-        return out._inplace_copy(result)
-
-    return lnstensor(result, from_lns=True, b=x.base)
-
-class LNSPositiveFunction(LNSFunction):
-    """
-    This is implemented solely for completeness, this
-    operation returns the input.
-
-    Gradients are calculated as follows:
-    d/dx(x) = 1
-    """
-
-    @staticmethod
-    def forward(x):
-        return x
-
-    @staticmethod
-    def setup_context(ctx, inputs, output):
-        pass # no context needed for this operation
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        return grad_output
-
-@implements(torch.positive, LNSPositiveFunction.forward, "default", default=True)
-def positive(x):
-
-    result = LNSPositiveFunction.apply(x)
-    return lnstensor(result, from_lns=True, b=x.base)
-
-class LNSSignFunction(LNSFunction):
-    """
-    Sign becomes checking the sign bit (rightmost bit).
-
-    Gradients are computed as follows:
-    d/dx(sign(x)) = 0
-    """
-
-    @staticmethod
-    def forward(x, base):
-        x_packed = x.to(torch.int64)
-        x_packed_sign = x_packed & 1
-
-        return torch.where(
-            torch.eq(x_packed | 1, LNS_ZERO), LNS_ZERO,
-            torch.where(x_packed_sign == 1,
-                        LNSTensor.get_internal_tensor(-1.0, base), LNSTensor.get_internal_tensor(1.0, base)))
-
-    @staticmethod
-    def setup_context(ctx, inputs, output):
-        pass # no context needed for this operation
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        return torch.full_like(grad_output, LNS_ZERO), None
-
-@implements(torch.sign, LNSSignFunction.forward, "default", default=True)
-def sign(x, *, out=None):
-
-    result = LNSSignFunction.apply(x, x.base)
 
     if out is not None:
         return out._inplace_copy(result)
